hdl: add can and should RX/TX
These variables control if an RX or TX transaction can occur (can_perform_*) due
to FIFO requirements being met. Should_perform_* determines if it is fair for
that particular operation to execute, ie. it is not fair to the RX queue if the
TX queue uses the GPIF bus twice in a row if the RX queue has data pending.
However, if there is no RX data pending, it is fair for the TX queue to use two
subsequent GPIF bus time slots.