fx3 firmware: enable 32bit GPIF
Due to an unconfirmed bug in the SDK's support for the GPIF II, it is not
xpossible to switch a 16bit GPIF over to 32bit funtionality once the FX3
enumerates. As a temporary work around, the firmware now uses a 32bit GPIF bus
when loading the Cyclone 4. This introduces some issues with headless execution
because enabling 32bit mode disables the SPI peripheral, which is how the FX3
communicates with the SPI flash. Conversely, in headless mode the unit has no
need to enter 32bit mode because it is not connected via USB3.0, therefore
during headless mode the bladeRF can choose to configure the GPIF 2 into 16bit
mode.