FX3 Firmware: Refactored build
These changes fix a regression caused by changeset
dc187e0640638b1895218299f47a4f392fdf05a7, which breaks the build
under Windows.

The firmware can now be built from the command line in both Linux and
Windows. The Eclipse project files now use the same Makefile as the
command line build, rather than generated Makefiles. As such, the
generated Debug/ has been removed.

Release builds are the default. To perform a debug build, run make
with DEBUG=y.

Additionally, the dependencies upon the FX3 toolchain makefiles have
been removed. The toolchain.mk.sample is now self-contained and
does not rely directly on environment variables. This is intended
to reduce the complexity of the build and aid in troubleshooting.

With the aforementioned changes, the compilation, linking and image
generation are intended to remain the same as previous changeset
(i.e., not affect the firmware). If anyone finds this *not* to be the
case, it should likely be regarded as a defect introduced by
this changeset.

This has been tested under the following configurations:
 - Linux, XUbuntu 13.04, FX3 SDK 1.2.2 w/ arm-2011.03 toolchain
 - Windows 7, FX3 SDK 1.2.3 (arm-2011.03-42 toolchain)

Linux Eclipse users will need to make a small change in their project
settings:
 - bladeRF -> properties -> C/C++ Build -> Builder Settings:
   + Uncheck "Use default build command" and change "cs-make" to "make"