fx3_firmware: disable multi DMA channel utilization for TX
Currently the FPGA tells the FX3 which of 4 (2 for RX and 2 for TX) "threads" to
use. These threads are actually hardware DMA engines. When multi DMAing occurs
for TX, 0-indexed DMA threads 2 and 3 are interleaved to copy data from the USB
to the FX3 GPIF port. However, there appears to be a bug in this interface
because as thread 2 is running, it will display the first word of thread 3
momentarily. The current solution is to downgrade to using a singel DMA channel,
which can easily handle upwards of 200MB/s of data per second; hence there
should be no performance hit.