libbladeRF: switch to RFLink mode if FPGA is programmed
This commit was introduced because the change over to using USB alternate
settings causes the libusb to change the device to its first alt-interface when
it is released. This basically means that as soon as the command line is closed
the device goes back to alt-interface 0. This commit figures out when this
happens and manually goes back to RFLink mode, alt-interface #1.