fx3_firmware: Add the ability to leave SPI/OTP programming mode
This commit allows the FX3 to exit SPI/OTP mode, and go back to the RFLink. This
transition wasn't allowed before because it was used to force the user to reboot
their unit after upgrading. It is now required because SPI flash/OTP commands
that read calibration and factory data can occur at anytime.