libbladeRF: add legacy mode support for devices running <v1.1
Lagecay mode support allows libbladeRF to upgrade and downgrade the firmware to
and from any version. All other features such as streaming samples, and talking
to any of the on-board devices will not work; the user has to upgrade the
firmware to the latest version before anything else can be done. The command
line now prints a message when legacy mode is enabled, and asks the user to
upgrade to the latest version.