bladeRF-cli: Reverted ee364bba3e2c and fixed return value issue(s)
The underlying problem for Issue #43 was a return of -1 for any
failures, rather than an apppropriate CMD_RET_* code. Since -1 ==
CMD_RET_QUIT, this error would cause the application to quit (hence the
issue being labeled a "crash")

The reworked parsing has been rolled back since the parsing was not the
source of the issue. Changes from ee364bba3e2c will be cherry-picked
back in as needed.