libbladeRF: Updated calibration routine.
Also updated the calibration routine in the CLI.  The CLI might need rethinking since, for the calibration to be successful, the module must be enabled.  Currently unless we're actively transmitting or receiving, the module is disabled.

Need to add a way to figure out if the module is enabled before trying to calibrate it.