libbladeRF: add FPGA flash autoload functionality
This commit adds the command line and libbladeRF support for flashing an image
to flash so that the bladeRF will auto-load the image when it boots. An rbf can
be loaded by running `bladeRF-cli -H ./hostedx40.rbf`. This will cause the FX3
firmware to load the FPGA image from flash the next time the unit boots, this
should not cause any problems reprogramming the device after it's already
loaded. To disable the auto-load feature after enabling it run
`bladeRF-cli -H X`, this destroys the FPGA image stored in flash.