bladeRF-cli: Improved script support, quoted param support, misc. cleanup
Scripts can be invoked from within other scripts, using the "run"
command. An error is reported if a recursive loop is detected.

An "echo" command has been added for debug purposes.

Parameters may now be surrounded in double quotes to include spaces.

Removed extraneous error messages from main(); The preceding functions
print error messages if needed.

Removed error-out if no device is attached. This allows scripts to be
run if the device is not attached. Commands already verify that a device
is attached, if required.