Cleanup FPGA autoloading.
Initialize the device if a default value is seen on the GPIO
config pins.

Fix up some coverity and samplerate floating point exception
bugs in the si5338 code.

Change the flash FPGA option to be -L instead of -H.

Make the version strings be long options only, no short ones.