Fixed minor issues.
ARST_LVL was inadvertantly changed to be a '0' instead of a '1' during the renaming of oc_i2c_master to bladerf_oc_i2c_master.  This is moved back.

Updated to using Quartus II 13.1, and with that discovered that Qsys now outputs '1' or '0' into a Verilog file which should be 1'b0 or 1'b1, so now there is a small sed replacement after generating the Qsys components to handle this problem.