host: Added option to disable "warnings as errors" in builds
Developers should always build with -Werror or /WX. However, we've had a
few cases with different compiler versions throwing errors that devs
didn't see, breaking master for users. The intention of this change is
to give those users a quick workaround, rather than requiring them to
hack around the warning in the source.