fx3_firmware: Consolidated and GPIF (re)configuration in gpif.c
The PIB and GPIF (re)configuration have been tucked away in gpif.c,
which is now used by the FPGA loader and RF link configurations.

Unnecessary PIB and IOMatrix configurations have been removed.

This fixes Issue #195, as the FPGA loader previously did not include a
PIB reset, which has been found to be required when changing from a
16-bit to 32-bit GPIF configuration.