libbladeRF: Rearranged and cleaned up header, removed unimplemented bladerf_stats
Definitions and typedefs in the libbladeRF.h header have been moved to
their appropriate groups so that they show up in the relevant module in
Doxygen output.

The bladerf_stats structure and associated function have been removed,
as this functionality is not schedule for implementation in the near
future. This will be re-added when this functionality becomes available.

The bladerf_set_loopback() function has been marked unimplemented as it
currently does not function as desired, but is schedule to be
implemented.