libbladeRF: Only clear sync buffer status for TX items
When starting an RX stream, the first N buffers, where N is the number
of transfers, will be submitted when the bladerf_stream() call is made.

Therefore, it is incorrect to clear IN_FLIGHT status; this will result in an
assertion firing in the RX callback.

After looking into this, it's become clear that the sync interface does
not correctly handle the situation where the underlying async stream
starts up. Issue #229 has been opened to further track the associated
defect and resolution.