hdl: prepare NiOS for expansion board upgrade
This commit adds a secondary select like to the DAC SPI controller (the
secondary select line is meant for the expansion board). Two parallel PIO ports
are are also added. One is an in/out PIO for expansion board GPIO values and the
other is for controlling the direction of the aforementioned PIO.