hdl: add FPGA based digital loopback RX mux
This commit introduces an RX mux mode where TX samples that are about to be
written to the LMS are piped into the RX sample FIFO. RX and TX must be running
at the sample sample rate. RX was not designed to latch on any specific event so
the first TX sample is dropped by RX FIFO. Software utilities looking to test
the behavior of the RX FIFO should anticipate that behavior as normal.