bladeRF-cli: Initialized value for help-interactive option flag
Previously, I hadn't explicitly initialized the flag used to indicate
whether the --help-interactive command line option had been provided,
which would sometimes cause the --help-interactive output to be
displayed even when the option was not specified. In particular, when
help2man would execute the --version option, it would sometimes instead
get the --help-interactive output, which was obviously no good.