libbladeRF: Removed end_fpga_programming() call from FPGA load
This function was causing intermittent failures because it would
return an error if it detected that that FPGA was not yet loaded,
which is a likely possibility.

Since this function no longer issues a vendor request to device to
perform a post FPGA load operation, it serves no purpose and has been
removed.