libbladeRF, bladeRF-cli: DC calibration table entry size fixes, added fields
The DC calibration table entries' offset values are int16_t, not
uint8_t. Although the actual register values are uint8_t, the libbladeRF
interface has normalized them.

We might want to bypass this normalization to reduce the size of this
calibration table, especially if we want to fit it in flash.