bladeRF-cli: Fixed invalid memset length when zero-padding a buffer
The memset used to zero the remainder of a buffer contained an incorrect
length calculation that could result in negative value, which is then
used as an unsigned value. This resulted in at attempt to zero of a very
large region of memory.

This calculation was failing to convert samples to int16_t's, and the
final length to zero (in int16_t's) to bytes.

Thank you to drmpeg and LazyDodo for their help on debugging and
resolving this issue.