libbladeRF: Fixed incorrect LMS LPF register configuration
Two issues existed in the LMS LPF configuration:

(1) The LPF bypass was not being disabled when switching to
BLADERF_LPF_DISABLED. Therefore, its state would be different depending
on whether the previous state was BLADERF_LPF_BYPASSED or
BLADERF_LPF_NORMAL.

(2) BLADERF_LPF_BYPASSED was leaving the LPF powered on. According to
the TX/RX LPF block diagram (Figure 3.5, in v1.1r4) in the LMS6002D
programming guide, this implies that the LPF was enabled while its
input was shorted to its output.