bladeRF-cli: Use synchronous API calls for rx/tx command implementations
As noted in the libbladeRF API documentation, long operations should not
be performed in the asynch stream callbacks. Given that the rx/tx
tasks are mainly performing disk I/O in these callbacks, the synchronous
interface is more appropriate.  Secondly, the sync interface is simpler,
lending itself to more readable code.

To further improve readability, the logic dealing with repetitions and
delays in the tx command has been changed to a state machine.

Additionally, the delays in the TX command are now implemented by
transmitting an appropriate number of zero samples. This yields more
accurate delays.