libbladeRF: Added temporary workaround for FX3 "Already started" error
This error from the FX3 is likely harmless, and I'll investigate further
soon. It only appears to occur when the firmware loopback mode times
out, and does not hinder further use of the device.

Given the lack of apparent severity, this band-aid is intended to keep
libbladeRF from overreacting until some time can be spend fixing the
underlying issue.