host: Removed bladeRF-flash
A majority of functionality presented by bladeRF-flash is now duplicated
in the bladeRF-cli utility.

Now that the bladeRF USB interfaces have stabilized more and the
instructions for bootloader-based recovery are available, this program
is generally no longer needed. As such, it is being removed to avoid
future neglect and bitrot.