bladeRF-cli: Fixed expr usage in generate.bash
expr does not appear to have a "match" operator in OSX. This has been
replaced with the ":" syntax for compatibility.

Additionally, a "--" argument has been added to avoid issues with
"$line" evaluating to an argument.

This fixes #278. Thank you to Mathieu Rene (@mrene) for reporting this
issue and providing these fixes.