host CMake: missing deps now FATAL_ERROR (#292)
In the event that BUILD_DOCUMENTATION is ON but Doxygen
or help2man are missing, or in the event ENABLE_LIBTECLA
is ON and libtecla is missing, raise a FATAL_ERROR and
abort CMake.

Since BUILD_DOCUMENTATION is OFF by default, and the existence
of libtecla determines the default for ENABLE_LIBTECLA, this
should have no impact for the casual "cmake .." user.