libbladeRF: mingw32 special case in file_ops.c
It's disgusting, it's terrible, and it's actually somewhat lewd,
but it causes the build to succeed.

Basically, SHGetKnownFolderPath should exist in libshell32.a
but it does not.  It'll probably work just fine on a Windows
system, but I can't link it as it is.

In theory, SHGetFolderPath may still exist.  Going to check
this out.

Fixes: #305