libbladeRF: Return "unsupported" when trying to enable the XB-100
This addresses issue #309. The XB-100 isn't currently supported via
libbladeRF. This support will come in the future, after the necessary
functionality is presented in the FPGA/firmware to expose this to the
host.

xb_attach() in xb.c was written in a way that would result in a
successful error code being reported when provided with an invalid or
unsupported expansion board type. This has been addressed, with some
cleanup to this function.

In making this change, it was noted that setting BLADERF_XB_NONE was
not disabling an attached expansion board.  A FIXME was added here to
draw attention to this; I've registered an issue for this and will
assign it to the author of the XB code.