cmake: Fixed 'make uninstall' issue (#314)
This addresses an error where 'make uninstall' would fail with:

    CMake Error at cmake_uninstall.cmake:4 (message):
        Cannot find install manifest:
        "<path>/build/host/install_manifest.txt".

The problem was that using CMAKE_CURRENT_BINARY_DIR in
cmake_uninstall.cmake wasn't really appropriate in our particular case,
as the "current" directory would be host/ once CMake dove into that
subdirectory.  This has been changed to CMAKE_BINARY_DIR to ensure
we're reference the top-level of the build.