libbladeRF: Check for BLADERF_FORCE_NO_FPGA_PRESENT environment variable
Setting this environment variable forces libbladeRF to always report
that no FPGA is loaded, rather than query the device.

This is intended for debugging and to provide a workaround for
issue #327, where users with a pre-v0.0.1 FPGA being autoloaded from SPI
flash are unable to open a device handle because the FPGA does not
supply a version number (which is now required by libbladeRF).