bladeRF-cli: Matched mimo cmd usage to help text, status code cleanup
The help text reports the usage of the mimo command to be:
    mimo [master | slave]

Since there was no alternative to argv[1] being anything other than "clk"
this was dropped so the invocation matches the help text.

If an error occurred, this command was returning a libbladeRF status
code directly, instead of returning a CLI_RET code. This has been fixed.