bladeRF-cli: Check if the device is streaming in main command handler
It was found that the 'calibrate' command could (incorrectly) be
executed while actively transmitting or receiving, which causes issues.

Since a number of commands already check if the device is currently
streaming, this check has been moved to the main command loop. Command
definitions now denote if they can be executed safely while the device
is streaming via an 'allow_while_streaming' flag.

The flash-related, open, and load commands have had their checks
removed, as they are now covered by the above flag.