libbladeRF: Ensure dev->cal.dc_<rx|tx> pointers are NULL after free
While reviewing config.c, it was noted that before attempting to
load a DC cal table, the old one (if present) is first freed.

While the current code doesn't provide the ability to unload and reload
DC calibration tables, this change is intended to avoid double free
issues when such functionality is added.