fx3_firmware: Added "device ready" query
This "device ready" status is used to denote that the device is
ready to be configured and used. Until the device becomes "ready"
all control requests will return an error.

This provides libbladeRF with a means to query when the device is done
loading an FPGA from flash, as well as any other firmware-based
initializations that may be needed in the future.