libbladeRF: Use Device recipient (in bmRequestType) for control transfers
The Interface recipient was being (mis)used. When specifying the Interface
recipient, wIndex[7:0] should be the desired interface number. WinUSB was
enforcing this, which was causing SPI flash page operations to be corrupted.

These requests have been changed to specify the Device recipient, which
does not appear to place restrictions on wIndex or wValue, allowing
us to use them as planned.