libbladeRF: Added FX3 firmware read and validation routines
This changeset introduces functionality for reading and validating an FX3
firmware file, as well as for accessing the firmware data in a
per-section manner.

The validation currently consists of testing the checksum, checking
size fields for bogus values, detecting a truncated file, and
ensuring that RAM addresses are in the SYSMEM or I-TCM regions of
the FX3's address space, assuming the default linker script is used
(which the bladeRF firmware does use).

This is based upon the boot image format documented in Cypress's AN76405:
"EZ-USB (R) FX3 (TM) Boot Options" (http://www.cypress.com/?docID=49862)