bladeRF-cli: Return non-zero code when --probe does not find devices
An optional `strict` parameter has been added to the interactive
mode `probe` command. Specifying this parameter instructs the command
to treat the case where no devices are found as an error.

Without specifying this `strict` parameter, the command will operate
as it has previously, treating this case as normal operation.

The handling of the --probe command-line option now uses `probe strict`,
as this provides shell scripts an easy way to determine if devices were
or were not found.