libbladeRF: Perform port reset when opening device (Linux, USB 3.0)
This changeset is directly based upon a patch and suggestions kindly
provided by Marian and Diana from Null Team (www.null.ro) and YateBTS
(www.yatebts.com). The original patch may be found on the issue tracker
for issue #95.

Per issue #95, it has been found that on many Linux-based
USB 3.0 setups, it is not possible to re-connect to a bladeRF after
a previous application using the bladeRF failed to close its device
handle. (For example, as the result of an application crashing or being
killed.)

To recover from this situation, we now perform a USB port reset when
opening the device, prior to attempting to actually access it.

Should this behavior cause any new issues, it can be disabled...

    ... at compile-time by providing
        -DENABLE_LIBUSB_DEV_RESET_ON_OPEN=False to CMake

    ... at run-time via the bladerf_set_usb_reset_on_open(false),
        when the aforementioned compile-time flag is True.

A huge "thank you" goes out Marian, Diana, and everyone at Null Team and
YateBTS!