libbladeRF: Updated LMS FREQSEL table to avoid PLL instability
Some of the changes in the table entry occur close to the end of
a VCO's freq range (per the plot in the LMS6002D FAQ, section 5.24).

As such, it was observed that the PLLs could lose lock over temperature
fluctuations.

The table entries have been updated and made more conservative with
respect where the VCO selection and/or divider ("FREQRANGE") change is
made.

Additionally, the table format has been changed to make it clear which
VCO and associated divider is being used.

A big thanks goes out to the Yate.ro team for their efforts in
identifying and resolving PLL instability issues.