libbladeRF: Added device retune mode selection
As of FPGA v0.2.0, the frequency tuning algorithm is run in the FPGA to
eliminate time wasted by communicating LMS6002D registers over USB. This
is now the default if the FPGA version is >= v0.2.0.

This default can be overridden via a BLADERF_DEFAULT_TUNE_MODE
environment variable, with options being "host" or "fpga".