lms: Rewrote LMS6002D Tuning algorithm
A VCOCAP hint is used as a starting point. The algorithm then walks
either up or down to first find the high or low limit, and then walks
the other way to find the other limit.

These seems to be working well, except we commonly run into a situation
where we never actually see VTUNE = LOW when our estimate is 54 or 55.